Refactored Stats bars
