+++ src-tauri/src/data.rs
@@ -9,3 +9,6 @@ pub use self::point::Point;
 
 mod point_cloud;
 pub use self::point_cloud::{point_selector, PointCloud};
+
+mod stats_info;
+pub use self::stats_info::{StatInfo, StatusBarConfig, StatusBarKind};
