+++ src-tauri/src/data/stats_info.rs
@@ -1,0 +1,179 @@
+use std::{fmt, time::Instant};
+
+use crate::image_analyzer::ImageAnalyzer;
+
+#[derive(Debug, Default, Clone, Copy)]
+pub enum StatusBarKind {
+    #[default]
+    Hp,
+    Mp,
+    Fp,
+    Xp,
+    EnemyHp,
+    SpellCasting,
+}
+impl fmt::Display for StatusBarKind {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            StatusBarKind::Hp => write!(f, "HP"),
+            StatusBarKind::Mp => write!(f, "MP"),
+            StatusBarKind::Fp => write!(f, "FP"),
+            StatusBarKind::Xp => write!(f, "XP"),
+            StatusBarKind::EnemyHp => write!(f, "enemy HP"),
+            StatusBarKind::SpellCasting => write!(f, "spell cast"),
+        }
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy)]
+pub struct StatInfo {
+    pub max_w: u32,
+    pub value: u32,
+    pub stat_kind: StatusBarKind,
+    pub last_update: Option<Instant>,
+}
+
+impl PartialEq for StatInfo {
+    fn eq(&self, other: &Self) -> bool {
+        self.value == other.value
+    }
+}
+
+impl PartialOrd for StatInfo {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.value.cmp(&other.value))
+    }
+}
+
+impl StatInfo {
+    pub fn new(
+        max_w: u32,
+        value: u32,
+        stat_kind: StatusBarKind,
+        image: Option<&ImageAnalyzer>,
+    ) -> Self {
+        let mut res = Self {
+            max_w,
+            value,
+            stat_kind,
+            last_update: Some(Instant::now()),
+        };
+        if image.is_some() {
+            res.update_value(image.unwrap());
+        }
+        res
+    }
+    pub fn update_value(&mut self, image: &ImageAnalyzer) {
+        let (updated_max_w, updated_value) = image.detect_status_bar(*self).unwrap_or_default();
+        let (old_max_w, old_value) = (self.max_w, self.value);
+        let mut changed = false;
+        if updated_max_w != old_max_w {
+            self.max_w = updated_max_w;
+            changed = true;
+        }
+        if updated_value != old_value {
+            self.value = updated_value;
+            changed = true;
+        }
+        if changed {
+            //println!("Update {} old value {} new {} last change was {}s", self.stat_kind.to_string(), old_value, self.value, self.last_update.unwrap().elapsed().as_secs());
+            self.last_update = Some(Instant::now());
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct StatusBarConfig {
+    pub max_search_x: u32,
+    pub max_search_y: u32,
+    pub min_search_x: u32,
+    pub min_search_y: u32,
+    pub refs: [[u8; 3]; 4],
+}
+
+impl StatusBarConfig {
+    pub fn new(colors: [[u8; 3]; 4]) -> Self {
+        Self {
+            refs: colors,
+            ..Default::default()
+        }
+    }
+}
+
+impl From<StatusBarKind> for StatusBarConfig {
+    fn from(kind: StatusBarKind) -> Self {
+        use StatusBarKind::*;
+
+        match kind {
+            Hp => {
+                StatusBarConfig::new([[174, 18, 55], [188, 24, 62], [204, 30, 70], [220, 36, 78]])
+            }
+
+            Mp => StatusBarConfig::new([
+                [20, 84, 196],
+                [36, 132, 220],
+                [44, 164, 228],
+                [56, 188, 232],
+            ]),
+            Fp => {
+                StatusBarConfig::new([[45, 230, 29], [28, 172, 28], [44, 124, 52], [20, 146, 20]])
+            }
+
+            Xp => StatusBarConfig::new([
+                [48, 185, 244],
+                [128, 212, 245],
+                [52, 196, 252],
+                [92, 236, 252],
+            ]),
+            EnemyHp => {
+                let mut enemy_hp_bar = StatusBarConfig::new([
+                    [174, 18, 55],
+                    [188, 24, 62],
+                    [204, 30, 70],
+                    [220, 36, 78],
+                ]);
+                enemy_hp_bar.min_search_x = 310;
+                enemy_hp_bar.min_search_y = 30;
+
+                enemy_hp_bar.max_search_x = 1000;
+                enemy_hp_bar.max_search_y = 60;
+
+                enemy_hp_bar
+            }
+            SpellCasting => {
+                let mut spell_casting_bar = StatusBarConfig::new([
+                    [16, 186, 15],
+                    [20, 157, 20],
+                    [15, 210, 14],
+                    [92, 164, 92],
+                ]);
+                spell_casting_bar.min_search_x = 310;
+                spell_casting_bar.min_search_y = 500;
+                // 800 -> 1038 fullscreen
+                //
+                spell_casting_bar.max_search_x = 1000;
+                spell_casting_bar.max_search_y = 1080;
+
+                spell_casting_bar
+            }
+        }
+    }
+}
+
+impl Default for StatusBarConfig {
+    fn default() -> Self {
+        Self {
+            max_search_x: 310,
+            max_search_y: 120,
+            min_search_x: 0,
+            min_search_y: 0,
+            refs: [[0; 3]; 4],
+        }
+    }
+}
+
+impl PartialEq for StatusBarConfig {
+    fn eq(&self, other: &Self) -> bool {
+        self.refs == other.refs && self.max_search_x == other.max_search_x
+    }
+}
